requirements.txt
Flask==3.1.0
Flask-WTF==1.2.2
numpy==2.1.3
pandas==2.2.3
pydantic==2.9.2
pymongo==4.10.1
python-dotenv==1.0.1
